Restablished map function
